requirements.txt
torch>=2.0.0
tensorflow>=2.13.0
transformers>=4.35.0
gradio>=4.0.0
librosa>=0.10.0
datasets>=2.14.0
accelerate>=0.24.0
safetensors>=0.4.0
soundfile>=0.12.0
numpy>=1.24.0
scipy>=1.11.0
python-dotenv>=1.0.0
pydub>=0.25.0
ffmpeg-python>=0.2.0
